// __inline-deps__/data-plane-gateway/src/gen/broker/protocol/broker.ts
export interface ProtocolLabel {
  name?: string;
  value?: string;
}
// ...
export interface ProtocolLabelSet {
  /** Labels of the set. Instances must be unique and sorted over (Name, Value). */
  labels?: ProtocolLabel[];
}